requirements.txt
ray>=2.10
dgl>=2.1
numpy
pydantic>=2
fastapi
uvicorn
grpcio
protobuf
pyyaml
psycopg2-binary
pgvector
neo4j-driver
asyncpg
tqdm
aiohttp
prometheus_client
